Arreglo en constructor

+++ src/Cliente.java
@@ -2,7 +2,7 @@ public class Cliente extends Thread {
     private long id;
     private String nombre;
 
-    public Cliente(int id, String nombre) {
+    public Cliente(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
